refactor server config
