+++ cmd/server/main.go
@@ -1,9 +1,8 @@
 package main
 
 import (
-	"flag"
 	"github.com/Carbohz/go-musthave-devops/internal/handler"
-	"github.com/caarlos0/env/v6"
+	"github.com/Carbohz/go-musthave-devops/internal/server"
 	"github.com/go-chi/chi"
 	"github.com/go-chi/chi/middleware"
 	"log"
@@ -12,19 +11,14 @@ import (
 	"os/signal"
 	"syscall"
 	"text/template"
-	"time"
 )
 
 const (
-	htmlFile 				= "index.html"
-	defaultAddress 			= "127.0.0.1:8080"
-	defaultStoreInterval 	= 300 * time.Second
-	defaultStoreFile 		= "/tmp/devops-metrics-db.json"
-	defaultRestore 			= true
+	htmlFile = "index.html"
 )
 
 func main() {
-	cfg := CreateConfig()
+	cfg := server.CreateConfig()
 	PrepareHTMLPage()
 	exitChan := make(chan int, 1)
 	go AwaitInterruptSignal(exitChan)
@@ -47,7 +41,7 @@ func PrepareHTMLPage() {
 	}
 }
 
-func RunServer(cfg handler.Config) {
+func RunServer(cfg server.Config) {
 	if cfg.Restore && cfg.StoreFile != "" {
 		handler.LoadMetrics(cfg)
 	}
@@ -66,45 +60,6 @@ func RunServer(cfg handler.Config) {
 	server.SetKeepAlivesEnabled(false)
 	log.Printf("Listening on address %s", cfg.Address)
 	log.Fatal(server.ListenAndServe())
-}
-
-func CreateConfig() handler.Config {
-	var cfg handler.Config
-	err := env.Parse(&cfg)
-	if err != nil {
-		log.Fatal(err)
-	}
-	log.Printf("Server is running with environment variables: %+v", cfg)
-
-	addressFlagPtr := flag.String("a", defaultAddress, "set address of server")
-	storeIntervalFlagPtr := flag.Duration("i", defaultStoreInterval, "set server's metrics store interval")
-	storeFileFlagPtr := flag.String("f", defaultStoreFile, "set file where metrics are stored")
-	restoreFlagPtr := flag.Bool("r", defaultRestore, "choose whether to restore server metrics from file")
-	flag.Parse()
-	log.Printf("Server is running with command line flags: Address %v, Store Interval %v, Store File %v, Restore %v",
-		*addressFlagPtr, *storeIntervalFlagPtr, *storeFileFlagPtr, *restoreFlagPtr)
-
-	_, isSet := os.LookupEnv("ADDRESS")
-	if !isSet {
-		cfg.Address = *addressFlagPtr
-	}
-
-	_, isSet = os.LookupEnv("STORE_INTERVAL")
-	if !isSet {
-		cfg.StoreInterval = *storeIntervalFlagPtr
-	}
-
-	_, isSet = os.LookupEnv("STORE_FILE")
-	if !isSet {
-		cfg.StoreFile = *storeFileFlagPtr
-	}
-
-	_, isSet = os.LookupEnv("RESTORE")
-	if !isSet {
-		cfg.Restore = *restoreFlagPtr
-	}
-
-	return cfg
 }
 
 func AwaitInterruptSignal(exitChan chan<- int) {
